working tables but main function broken
