+++ src/models/species.py
@@ -1,0 +1,13 @@
+from datetime import datetime
+from sqlalchemy import create_engine, Column, Integer, String, ForeignKey, Date
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import relationship, sessionmaker
+from src.models.base import Base
+
+class Species(Base):
+    __tablename__ = 'species'
+    id = Column(Integer, primary_key=True)
+    name = Column(String(50))
+    life_span = Column(Integer)
+
+    animals = relationship('Animal', back_populates='species')
